requirements.txt
PyPDF2
pytesseract
#python-docx
pillow
python-dotenv
google-generativeai
pinecone
sentence-transformers

